requirement.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
django-jalali-date==2.0.0
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
dotenv==0.9.9
drf-yasg==1.21.14
easy-thumbnails==2.10.1
gunicorn==25.0.3
idna==3.11
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
kavenegar==1.1.2
packaging==26.0
phonenumbers==9.0.23
pillow==12.1.1
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
django-filter